+++ euhat/common/base64.h
@@ -4,11 +4,11 @@
 extern "C" {
 #endif
 
-int base64EncodeLen(int len);
-int base64Encode(char *coded_dst, const char *plain_src, int len_plain_src);
+size_t base64EncodeLen(size_t len);
+size_t base64Encode(char *coded_dst, const char *plain_src, size_t len_plain_src);
 
-int base64DecodeLen(const char *coded_src);
-int Base64Decode(char *plain_dst, const char *coded_src);
+size_t base64DecodeLen(const char *coded_src);
+size_t Base64Decode(char *plain_dst, const char *coded_src);
 
 #ifdef __cplusplus
 }
